--- database/.env
@@ -1,3 +1,0 @@
-POSTGRES_USER=admin
-POSTGRES_PASSWORD=password
-POSTGRES_DB=postgres
