--- database/docker-compose.yml
@@ -1,13 +1,0 @@
-services:
-  db:
-    image: postgres:latest
-    container_name: my_postgres
-    restart: always
-    env_file:
-      .env
-    volumes:
-      - ./data:/var/lib/postgresql/data  # Persist data
-      - ./DDL.sql:/docker-entrypoint-initdb.d/DDL.sql
-    ports:
-      - "5432:5432"
-
