+++ project-api/src/main/kotlin/db/user/UserTable.kt
@@ -7,7 +7,6 @@ object UserTable : IntIdTable("user") {
     val name = varchar("name", 255)
     val email = varchar("email", 255)
     val password = varchar("hashpasswordsalted", 255)
-    val salt = varchar("salt", 128)  // Changed from 128 to 16 to match DB
-    val createdAt = timestamp("createdat").defaultExpression(CurrentTimestamp)  // Changed from varchar to timestamp
-
+    val salt = varchar("salt", 128)
+    val createdAt = timestamp("createdat").defaultExpression(CurrentTimestamp)
 }
