+++ project-api/src/main/resources/user.html
@@ -1,0 +1,156 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>User Profile</title>
+    <style>
+        body {
+            font-family: 'Segoe UI', Tahoma, Geneva, Verdana, sans-serif;
+            max-width: 800px;
+            margin: 0 auto;
+            padding: 20px;
+            color: #333;
+        }
+        .card {
+            background-color: #f9f9f9;
+            border-radius: 8px;
+            box-shadow: 0 2px 10px rgba(0, 0, 0, 0.1);
+            padding: 20px;
+            margin-top: 20px;
+        }
+        h1 {
+            color: #2c3e50;
+            margin-bottom: 30px;
+        }
+        .user-detail {
+            margin-bottom: 12px;
+            display: flex;
+        }
+        .detail-label {
+            font-weight: bold;
+            width: 150px;
+        }
+        .detail-value {
+            flex: 1;
+        }
+        .error-message {
+            color: #e74c3c;
+            padding: 10px;
+            border-left: 4px solid #e74c3c;
+            background-color: #fadbd8;
+            margin-top: 20px;
+        }
+        .loading {
+            text-align: center;
+            margin-top: 30px;
+            color: #7f8c8d;
+        }
+        button {
+            background-color: #3498db;
+            color: white;
+            border: none;
+            padding: 10px 15px;
+            border-radius: 4px;
+            cursor: pointer;
+            margin-top: 20px;
+        }
+        button:hover {
+            background-color: #2980b9;
+        }
+    </style>
+</head>
+<body>
+<h1>User Profile</h1>
+
+<div id="loading" class="loading">Loading user data...</div>
+<div id="error" class="error-message" style="display: none;"></div>
+<div id="user-container" class="card" style="display: none;">
+    <div class="user-detail">
+        <div class="detail-label">ID:</div>
+        <div id="user-id" class="detail-value"></div>
+    </div>
+    <div class="user-detail">
+        <div class="detail-label">Name:</div>
+        <div id="user-name" class="detail-value"></div>
+    </div>
+    <div class="user-detail">
+        <div class="detail-label">Email:</div>
+        <div id="user-email" class="detail-value"></div>
+    </div>
+    <div class="user-detail">
+        <div class="detail-label">Account Created:</div>
+        <div id="user-created" class="detail-value"></div>
+    </div>
+</div>
+
+<button id="refresh-btn" style="display: none;">Refresh User Data</button>
+
+<script>
+    document.addEventListener('DOMContentLoaded', function() {
+        fetchUserData();
+
+        document.getElementById('refresh-btn').addEventListener('click', function() {
+            fetchUserData();
+        });
+    });
+
+    function fetchUserData() {
+        // Show loading, hide other elements
+        document.getElementById('loading').style.display = 'block';
+        document.getElementById('error').style.display = 'none';
+        document.getElementById('user-container').style.display = 'none';
+
+        // Get token from local storage
+        const token = localStorage.getItem('token');
+
+        if (!token) {
+            showError('Authentication token not found. Please log in again.');
+            return;
+        }
+
+        // Make API request with the bearer token
+        fetch('/user', {
+            method: 'POST',
+            headers: {
+                'Authorization': `Bearer ${token}`,
+                'Content-Type': 'application/json'
+            }
+        })
+            .then(response => {
+                if (!response.ok) {
+                    throw new Error(`Error: ${response.status} - ${response.statusText}`);
+                }
+                return response.json();
+            })
+            .then(user => {
+                displayUserData(user);
+                document.getElementById('loading').style.display = 'none';
+                document.getElementById('user-container').style.display = 'block';
+                document.getElementById('refresh-btn').style.display = 'block';
+            })
+            .catch(error => {
+                console.error('Error fetching user data:', error);
+                showError(error.message || 'Failed to load user data. Please try again later.');
+            });
+    }
+
+    function displayUserData(user) {
+        document.getElementById('user-id').textContent = user.id;
+        document.getElementById('user-name').textContent = user.name;
+        document.getElementById('user-email').textContent = user.email;
+
+        const createdDate = new Date(user.createdAt);
+        document.getElementById('user-created').textContent = createdDate.toLocaleString();
+    }
+
+    function showError(message) {
+        const errorElement = document.getElementById('error');
+        errorElement.textContent = message;
+        errorElement.style.display = 'block';
+        document.getElementById('loading').style.display = 'none';
+        document.getElementById('refresh-btn').style.display = 'block';
+    }
+</script>
+</body>
+</html>
