+++ project-api/src/test/kotlin/routes/UsersKtTest.kt
@@ -7,7 +7,7 @@ import io.ktor.server.testing.*
 import kotlinx.serialization.json.Json
 import mx.unam.fciencias.ids.eq1.model.user.User
 import mx.unam.fciencias.ids.eq1.model.user.repository.UserRepository
-import mx.unam.fciencias.ids.eq1.routes.configureUsers
+import mx.unam.fciencias.ids.eq1.routes.users.configureUsers
 import mx.unam.fciencias.ids.eq1.service.users.DBUserService
 import mx.unam.fciencias.ids.eq1.service.users.UserService
 import org.junit.jupiter.api.AfterEach
